🔨[구현] 컨트롤러, 서비스, 리포지토리 분리
