+++ src/main/java/supercoding/mall/controller/UserController.java
@@ -22,23 +22,5 @@ public class UserController {
         userService.addUser(userDTO);
         return ResponseEntity.ok("회원가입 성공");
     }
-    @GetMapping("/view_products")
-    public ResponseEntity< Map<String, Map<String, Product>>> viewProducts(){
-        Map<String, Map<String, Product>> productMap = userService.viewProducts();
-        return ResponseEntity.ok(productMap);
-    }
-
-    @PostMapping("{id}/add")
-    public ResponseEntity<String> addProduct(@RequestBody ProductDTO productDTO,@PathVariable String id){
-        userService.addProduct(productDTO,id);
-        return ResponseEntity.ok("상품 추가 성공");
-    }
-
-    @PostMapping("{id}/delete/{product_id}")
-    public ResponseEntity<String> deleteProduct(@PathVariable String id,@PathVariable String product_id){
-        userService.deleteProduct(id,product_id);
-        return ResponseEntity.ok("상품 삭제 성공 상품번호:"+product_id);
-    }
-
 
 }
