+++ src/main/java/supercoding/mall/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package supercoding.mall.repository;
 
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import supercoding.mall.domain.Product;
 import supercoding.mall.domain.ProductDTO;
@@ -11,10 +12,10 @@ import java.util.*;
 
 @Repository
 @Getter
+@RequiredArgsConstructor
 public class UserRepository {
-
-    private Map<String,User> userMap = new HashMap<>();
-    private Map<String, Map<String, Product>> productMap = new HashMap<>();
+    @Getter
+    private static Map<String,User> userMap = new HashMap<>();
 
     public void addUser(int serialId,UserDTO userDTO){
         User user = User.builder()
@@ -26,37 +27,10 @@ public class UserRepository {
                 .build();
         userMap.put(user.getId(), user);
     }
-    public void addProduct(int serialId, ProductDTO productDTO, String userId) {
-        User findedUser = userMap.get(userId);
-        Product product = Product.builder()
-                .id(String.valueOf(serialId))
-                .productName(productDTO.getProductName())
-                .userId(findedUser.getUserId())
-                .price(productDTO.getPrice())
-                .quantity(productDTO.getQuantity())
-                .totalPrice(productDTO.getPrice() * productDTO.getQuantity())
-                .category(productDTO.getCategory().getName()).build();
-
-//        Map<String, Product> innerMap = productMap.computeIfAbsent(findedUser.getUserId(), k -> new HashMap<>());
-        Map<String, Product> innerMap = productMap.get(findedUser.getId());
-        if (innerMap == null) {
-            innerMap = new HashMap<>();
-            productMap.put(findedUser.getId(), innerMap);
-        }
-        innerMap.put(product.getId(), product);
-    }
 
     public long findByUserId(String userId){
         return userMap.entrySet().stream().filter(user->user.getValue().getUserId().equals(userId)).count();
     }
-    public Product findProduct(String userId, String productId) {
-        User findUser = userMap.get(userId);
-        Map<String, Product> products = productMap.get(findUser.getId());
-        return products.get(productId);
-    }
 
-    public void deleteProduct(String userId, String productId) {
-        productMap.get(userId).remove(productId);
-    }
 
 }
