+++ src/main/java/supercoding/mall/service/UserService.java
@@ -3,20 +3,15 @@ package supercoding.mall.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import supercoding.mall.domain.Product;
-import supercoding.mall.domain.ProductDTO;
-import supercoding.mall.domain.User;
 import supercoding.mall.domain.UserDTO;
 import supercoding.mall.repository.UserRepository;
-
-import java.util.Map;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private static int serialUserId = 1;
-    private static int serialProductId = 1;
+
     private final UserRepository userRepository;
 
     public void addUser(UserDTO userDTO){
@@ -28,16 +23,5 @@ public class UserService {
         }
     }
 
-    public void addProduct(ProductDTO productDTO,String userId){
-        userRepository.addProduct(serialProductId++ , productDTO, userId);
-    }
 
-    public void deleteProduct(String userId,String productId){
-        userRepository.deleteProduct(userId,productId);
-        serialProductId--;
-    }
-
-    public Map<String, Map<String, Product>> viewProducts(){
-        return userRepository.getProductMap();
-    }
 }
